Start the week with Monday,
because that's the localization in Europe, that's what I use personally.
I don't want to put a setting on it, because it would clutter the interface and
this only need to change once. The other two (color and sound) might want
somebody to change day-by-day, it's not a problem they are present at the bottom.
Fixes #1.